requirements.txt
moviepy
openai-whisper
openai
python-dotenv 